Bring web-frontend and psss e2e tests closer
