--- packages/psss/cypress.env.example.json
@@ -1,3 +1,0 @@
-{
-  "testUserPassword": ""
-}
