+++ packages/psss/cypress/plugins/index.js
@@ -1,1 +1,8 @@
-module.exports = (on, config) => {};
+/**
+ * Tupaia
+ * Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
+ */
+
+const cypressDotenv = require('cypress-dotenv');
+
+module.exports = (on, config) => cypressDotenv(config);
