+++ packages/web-frontend/cypress/plugins/index.js
@@ -6,9 +6,6 @@
 const cypressDotenv = require('cypress-dotenv');
 const fs = require('fs');
 
-/**
- * @type {Cypress.PluginConfig}
- */
 module.exports = (on, config) => {
   // `on` is used to hook into various events Cypress emits
   // `config` is the resolved Cypress config
